cleaning up model structures and removing unecessary utilfile
