+++ mtg/ml/models.py
@@ -1,112 +1,33 @@
 import tensorflow as tf
-from mtg.utils.display import build_decks
-from tensorflow.python.keras.engine.base_layer import Layer
 from mtg.ml import nn
-from mtg.ml.layers import MultiHeadAttention, Dense, LayerNormalization, Embedding
+from mtg.ml.layers import Dense, Embedding
 import numpy as np
-import pandas as pd
-import pdb
 import pathlib
 import os
 import pickle
-
-
-class CustomSchedule(tf.keras.optimizers.schedules.LearningRateSchedule):
-    def __init__(self, d_model, warmup_steps=1000):
-        super(CustomSchedule, self).__init__()
-
-        self.d_model = d_model
-        self.d_model = tf.cast(self.d_model, tf.float32)
-
-        self.warmup_steps = warmup_steps
-
-    def __call__(self, step):
-        arg1 = tf.math.rsqrt(step)
-        arg2 = step * (self.warmup_steps ** -1.5)
-
-        return tf.math.rsqrt(self.d_model) * tf.math.minimum(arg1, arg2)
-
-
-class ConcatEmbedding(tf.Module):
-    """
-    Lets say you want an embedding that is a concatenation of the abstract object and data about the object
-
-    so we learn a normal one hot embedding, and then have an MLP process the data about the object and concatenate the two.
-    """
-
-    def __init__(
-        self,
-        num_items,
-        emb_dim,
-        item_data,
-        dropout=0.0,
-        n_h_layers=1,
-        initializer=tf.initializers.GlorotNormal(),
-        name=None,
-        activation=None,
-        start_act=None,
-        middle_act=None,
-        out_act=None,
-    ):
-        super().__init__(name=name)
-        assert item_data.shape[0] == num_items
-        self.item_data = item_data
-        self.item_MLP = nn.MLP(
-            in_dim=item_data.shape[1],
-            start_dim=item_data.shape[1] // 2,
-            out_dim=emb_dim // 2,
-            n_h_layers=n_h_layers,
-            dropout=dropout,
-            name="item_data_mlp",
-            start_act=start_act,
-            middle_act=middle_act,
-            out_act=out_act,
-            style="bottleneck",
-        )
-        self.embedding = tf.Variable(
-            initializer(shape=(num_items, emb_dim // 2)),
-            dtype=tf.float32,
-            name=self.name + "_embedding",
-        )
-        self.activation = activation
-
-    @tf.function
-    def __call__(self, x, training=None):
-        # x is a card id
-        item_embeddings = tf.gather(self.embedding, x)
-        data_embeddings = tf.gather(
-            self.item_MLP(self.item_data, training=training), x,
-        )
-        embeddings = tf.concat([item_embeddings, data_embeddings], axis=-1)
-        if self.activation is not None:
-            embeddings = self.activation(embeddings)
-        return embeddings
+from mtg.ml.utils import CustomSchedule
 
 
 class DraftBot(tf.Module):
     def __init__(
         self,
-        cards,
+        expansion,
         emb_dim,
-        t,
         num_encoder_heads,
         num_decoder_heads,
         num_encoder_layers,
         num_decoder_layers,
         pointwise_ffn_width,
-        card_data=None,
         emb_dropout=0.0,
         memory_dropout=0.0,
         out_dropout=0.0,
-        use_deckbuilder=False,
-        output_MLP=False,
         name=None,
     ):
         super().__init__(name=name)
-        self.idx_to_name = cards.set_index("idx")["name"].to_dict()
-        self.card_data = cards
+        self.idx_to_name = expansion.create_mapping("idx", "name", include_basics=False)
         self.n_cards = len(self.idx_to_name)
-        self.t = t
+        self.t = expansion.t
+        self.card_data = expansion.card_data_for_ML[5:]
         self.emb_dim = tf.Variable(
             emb_dim, dtype=tf.float32, trainable=False, name="emb_dim"
         )
@@ -114,7 +35,7 @@ class DraftBot(tf.Module):
         self.positional_embedding = Embedding(t, emb_dim, name="positional_embedding")
         self.positional_mask = 1 - tf.linalg.band_part(tf.ones((t, t)), -1, 0)
         self.encoder_layers = [
-            TransformerBlock(
+            nn.TransformerBlock(
                 emb_dim,
                 num_encoder_heads,
                 pointwise_ffn_width,
@@ -125,22 +46,15 @@ class DraftBot(tf.Module):
         ]
         # extra embedding as representation of bias before the draft starts. This is grabbed as the
         # representation for the "previous pick" that goes into the decoder for P1P1
-        self.card_data = card_data
-        if self.card_data is None:
-            self.card_embedding = Embedding(
-                self.n_cards + 1, emb_dim, name="card_embedding", activation=None
-            )
-        else:
-            self.card_data = tf.convert_to_tensor(self.card_data, dtype=tf.float32)
-            self.card_embedding = ConcatEmbedding(
-                self.n_cards + 1,
-                emb_dim,
-                self.card_data,
-                name="card_embedding",
-                activation=None,
-            )
+        self.card_embedding = nn.ConcatEmbedding(
+            self.n_cards + 1,
+            emb_dim,
+            tf.convert_to_tensor(self.card_data, dtype=tf.float32),
+            name="card_embedding",
+            activation=None,
+        )
         self.decoder_layers = [
-            TransformerBlock(
+            nn.TransformerBlock(
                 emb_dim,
                 num_decoder_heads,
                 pointwise_ffn_width,
@@ -150,115 +64,63 @@ class DraftBot(tf.Module):
             )
             for i in range(num_decoder_layers)
         ]
-        self.output_MLP = output_MLP
-        if self.output_MLP:
-            self.output_decoder = nn.MLP(
-                in_dim=emb_dim,
-                start_dim=emb_dim * 2,
-                out_dim=self.n_cards,
-                n_h_layers=1,
-                dropout=out_dropout,
-                name="output_decoder",
-                start_act=tf.nn.selu,
-                middle_act=tf.nn.selu,
-                out_act=None,
-                style="reverse_bottleneck",
-            )
-        else:
-            self.output_decoder = Dense(
-                emb_dim, emb_dim, name="output_decoder", activation=None
-            )
-        if use_deckbuilder:
-            self.deckbuilder = DeckBuilder()
-        else:
-            self.deckbuilder = None
-        # initializer=tf.initializers.GlorotNormal()
-        # self.initial_card_bias = tf.Variable(
-        #     initializer(shape=(1, emb_dim)),
-        #     dtype=tf.float32,
-        #     name=self.name + "_initial_card_bias",
-        # )
+        self.output_decoder = nn.MLP(
+            in_dim=emb_dim,
+            start_dim=emb_dim * 2,
+            out_dim=self.n_cards,
+            n_h_layers=1,
+            dropout=out_dropout,
+            name="output_decoder",
+            start_act=tf.nn.selu,
+            middle_act=tf.nn.selu,
+            out_act=None,
+            style="reverse_bottleneck",
+        )
 
     @tf.function
     def __call__(
-        self, features, training=None, return_attention=False, return_build=True
+        self, features, training=None, return_attention=False,
     ):
-        if self.deckbuilder is not None and return_build:
-            packs, picks, positions, final_pools = features
-        else:
-            packs, picks, positions = features
-        # pools = draft_info[:, :, self.n_cards:]
-        # draft_info is of shape (batch_size, t, n_cards * 2)
+        packs, picks, positions = features
         positional_masks = tf.gather(self.positional_mask, positions)
         positional_embeddings = self.positional_embedding(positions, training=training)
-        # old way: pack embedding = mean of card embeddings for only cards in the pack
-        # batch_size x t x n_cards x emb_dim
         all_card_embeddings = self.card_embedding(
             tf.range(self.n_cards), training=training
         )
-        if self.deckbuilder is not None and return_build:
-            final_pool_embeddings = (
-                final_pools[:, :, None] * all_card_embeddings[None, :, :]
-            )
-            built_decks = self.deckbuilder(final_pool_embeddings, training=training)
         pack_card_embeddings = (
             packs[:, :, :, None] * all_card_embeddings[None, None, :, :]
         )
         n_options = tf.reduce_sum(packs, axis=-1, keepdims=True)
         pack_embeddings = tf.reduce_sum(pack_card_embeddings, axis=2) / n_options
         embs = pack_embeddings * tf.math.sqrt(self.emb_dim) + positional_embeddings
-        # insert an embedding to represent bias towards cards/archetypes/concepts you have before the draft starts
-        # --> this could range from "generic pick order of all cards" to "blue is the best color", etc etc
-        # batch_bias = tf.tile(tf.expand_dims(self.initial_card_bias,0), [embs.shape[0],1,1])
-        # batch_mask = tf.tile(tf.expand_dims(self.positional_mask,0), [embs.shape[0],1,1])
-        # embs = tf.concat([
-        #     batch_bias,
-        #     embs,
-        # ], axis=1)
+
         if training and self.dropout > 0.0:
             embs = tf.nn.dropout(embs, rate=self.dropout)
+
         for memory_layer in self.encoder_layers:
             embs, attention_weights_pack = memory_layer(
                 embs, positional_masks, training=training
             )  # (batch_size, t, emb_dim)
+
         dec_embs = self.card_embedding(picks, training=training)
         for memory_layer in self.decoder_layers:
             dec_embs, attention_weights_pick = memory_layer(
                 dec_embs, positional_masks, encoder_output=embs, training=training
             )  # (batch_size, t, emb_dim)
-        # dec_embs = dec_embs
-        # get rid of output with respect to initial bias vector, as that is not part of prediction
-        # embs = embs[:,1:,:]
+
         mask_for_softmax = 1e9 * (1 - packs)
-        if self.output_MLP:
-            card_rankings = (
-                self.output_decoder(dec_embs, training=training) * packs
-                - mask_for_softmax
-            )  # (batch_size, t, n_cards)
-            emb_dists = (
-                tf.sqrt(
-                    tf.reduce_sum(
-                        tf.square(pack_card_embeddings - dec_embs[:, :, None, :]), -1
-                    )
+        card_rankings = (
+            self.output_decoder(dec_embs, training=training) * packs - mask_for_softmax
+        )  # (batch_size, t, n_cards)
+        emb_dists = (
+            tf.sqrt(
+                tf.reduce_sum(
+                    tf.square(pack_card_embeddings - dec_embs[:, :, None, :]), -1
                 )
-                * packs
-                + mask_for_softmax
             )
-        else:
-            approx_pick_embs = self.output_decoder(dec_embs, training=training)
-            emb_dists = (
-                tf.sqrt(
-                    tf.reduce_sum(
-                        tf.square(
-                            pack_card_embeddings - approx_pick_embs[:, :, None, :]
-                        ),
-                        -1,
-                    )
-                )
-                * packs
-                + mask_for_softmax
-            )
-            card_rankings = -emb_dists
+            * packs
+            + mask_for_softmax
+        )
         output = tf.nn.softmax(card_rankings)
         # zero out the rankings for cards not in the pack
         # note1: this only works because no foils on arena means packs can never have 2x of a card
@@ -266,16 +128,7 @@ class DraftBot(tf.Module):
         # note2: this zeros out the gradients for the cards not in the pack in order to not negatively
         #        affect backprop on cards that would techncally be taken if they were in the pack. However,
         #        if it turns out that there is a reason why these gradients shouldn't be zero, this
-        #        multiplication could be done only during inference (when training is not True)
-
-        # add epsilon for cards in the pack to ensure they are non-zero (handles edge cases)
-        # card_rankings = card_rankings * packs + 1e-9 * packs
-        # after zeroing out cards not in packs, we readjust the output to maintain that it sums to one
-        # note: currently this sums to one so we do from_logits=True in Categorical Cross Entropy,
-        #       possible softmax is better than relu, regardless this does have numerical instability issues
-        #       so that is something to look out for. But from_logits=False had terrible performance
-        if self.deckbuilder is not None and return_build:
-            output = (output, built_decks)
+        #        multiplication could be done only during inference (when training is False)
         if return_attention:
             return output, (attention_weights_pack, attention_weights_pick)
         return output, emb_dists
@@ -291,7 +144,6 @@ class DraftBot(tf.Module):
         rare_lambda=1.0,
         cmc_lambda=1.0,
         cmc_margin=1.0,
-        card_data=None,
         metric_names=["top1", "top2", "top3"],
     ):
         if optimizer is None:
@@ -315,8 +167,7 @@ class DraftBot(tf.Module):
         self.rare_lambda = rare_lambda
         self.cmc_lambda = cmc_lambda
         self.cmc_margin = cmc_margin
-        if card_data is not None:
-            self.set_card_params(card_data)
+        self.set_card_params(self.card_data.iloc[:-1, :])
         self.metric_names = metric_names
 
     def set_card_params(self, card_data):
@@ -407,87 +258,6 @@ class DraftBot(tf.Module):
                 ),
             }
             pickle.dump(attrs, f)
-
-
-class TransformerBlock(tf.Module):
-    """
-    self attention block for encorporating memory into the draft bot
-    """
-
-    def __init__(
-        self, emb_dim, num_heads, pointwise_ffn_width, dropout=0.0, decode=False, name=None
-    ):
-        super().__init__(name=name)
-        self.dropout = dropout
-        self.decode = decode
-        # kdim and dmodel are the same because the embedding dimension of the non-attended
-        # embeddings are the same as the attention embeddings.
-        self.attention = MultiHeadAttention(
-            emb_dim, emb_dim, num_heads, name=self.name + "_attention"
-        )
-        self.expand_attention = Dense(
-            emb_dim,
-            pointwise_ffn_width,
-            activation=tf.nn.relu,
-            name=self.name + "_pointwise_in",
-        )
-        self.compress_expansion = Dense(
-            pointwise_ffn_width, emb_dim, activation=None, name=self.name + "_pointwise_out"
-        )
-        self.final_layer_norm = LayerNormalization(
-            emb_dim, name=self.name + "_out_norm"
-        )
-        self.attention_layer_norm = LayerNormalization(
-            emb_dim, name=self.name + "_attention_norm"
-        )
-        if self.decode:
-            self.decode_attention = MultiHeadAttention(
-                emb_dim, emb_dim, num_heads, name=self.name + "_decode_attention"
-            )
-            self.decode_layer_norm = LayerNormalization(
-                emb_dim, name=self.name + "_decode_norm"
-            )
-
-    def pointwise_fnn(self, x, training=None):
-        x = self.expand_attention(x, training=training)
-        return self.compress_expansion(x, training=training)
-
-    @tf.function
-    def __call__(self, x, mask, encoder_output=None, training=None):
-        attention_emb, attention_weights = self.attention(
-            x, x, x, mask, training=training
-        )
-        if training and self.dropout > 0:
-            attention_emb = tf.nn.dropout(attention_emb, rate=self.dropout)
-        residual_emb_w_memory = self.attention_layer_norm(
-            x + attention_emb, training=training
-        )
-        if self.decode:
-            assert encoder_output is not None
-            decode_attention_emb, decode_attention_weights = self.decode_attention(
-                encoder_output,
-                encoder_output,
-                residual_emb_w_memory,
-                mask,
-                training=training,
-            )
-            if training and self.dropout > 0:
-                decode_attention_emb = tf.nn.dropout(
-                    decode_attention_emb, rate=self.dropout
-                )
-            residual_emb_w_memory = self.decode_layer_norm(
-                residual_emb_w_memory + decode_attention_emb, training=training
-            )
-            attention_weights = (attention_weights, decode_attention_weights)
-        process_emb = self.pointwise_fnn(residual_emb_w_memory, training=training)
-        if training and self.dropout > 0:
-            process_emb = tf.nn.dropout(process_emb, rate=self.dropout)
-        return (
-            self.final_layer_norm(
-                residual_emb_w_memory + process_emb, training=training
-            ),
-            attention_weights,
-        )
 
 
 class DeckBuilder(tf.Module):
