+++ mtg/ml/utils.py
@@ -4,6 +4,26 @@ import pickle
 import os
 
 
+class CustomSchedule(tf.keras.optimizers.schedules.LearningRateSchedule):
+    """
+    learning rate scheduling
+    """
+
+    def __init__(self, d_model, warmup_steps=1000):
+        super(CustomSchedule, self).__init__()
+
+        self.d_model = d_model
+        self.d_model = tf.cast(self.d_model, tf.float32)
+
+        self.warmup_steps = warmup_steps
+
+    def __call__(self, step):
+        arg1 = tf.math.rsqrt(step)
+        arg2 = step * (self.warmup_steps ** -1.5)
+
+        return tf.math.rsqrt(self.d_model) * tf.math.minimum(arg1, arg2)
+
+
 def importance_weighting(df, minim=0.1, maxim=1.0):
     rank_to_score = {
         "bronze": 0.01,
